refactor(architecture): refactor uploads
