+++ consts/maxNumberOfFiles.js
@@ -1,0 +1,5 @@
+const MAX_NUMBER_OF_FILES = 5;
+
+module.exports = {
+ MAX_NUMBER_OF_FILES,
+};
