+++ router/uploads/swagger/deletePhoto.swagger.js
@@ -1,0 +1,24 @@
+/**
+ * @swagger
+ * /uploads/delete-photo:
+ *   delete:
+ *     summary: Delete a photo from S3 and database
+ *     tags: [Uploads]
+ *     security:
+ *       - bearerAuth: []
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               url:
+ *                 type: string
+ *                 example: "uploads/your-photo.jpg"
+ *     responses:
+ *       200:
+ *         description: Photo deleted successfully
+ *       404:
+ *         description: Photo not found
+ */
