+++ router/uploads/swagger/getLatestPhotos.swagger.js
@@ -1,0 +1,12 @@
+/**
+ * @swagger
+ * /uploads/latest:
+ *   get:
+ *     summary: Get latest 3 uploads
+ *     tags: [Uploads]
+ *     security:
+ *       - bearerAuth: []
+ *     responses:
+ *       200:
+ *         description: List of recent uploads
+ */
