+++ router/uploads/swagger/uploadMessagePhotos.swagger.js
@@ -1,0 +1,24 @@
+/**
+ * @swagger
+ * /uploads/message-photos:
+ *   post:
+ *     summary: Upload message-related photos
+ *     tags: [Uploads]
+ *     security:
+ *       - bearerAuth: []
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         multipart/form-data:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               avatars:
+ *                 type: array
+ *                 items:
+ *                   type: string
+ *                   format: binary
+ *     responses:
+ *       200:
+ *         description: Upload successful
+ */
